feat: improve TaskDetailsView
