--- packages/apps/project/src/components/atoms/TaskButton/TaskButton.tsx
@@ -1,36 +1,0 @@
-/*
- * Axelor Business Solutions
- *
- * Copyright (C) 2024 Axelor (<http://axelor.com>).
- *
- * This program is free software: you can redistribute it and/or  modify
- * it under the terms of the GNU Affero General Public License, version 3,
- * as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-import React from 'react';
-import {useTranslator} from '@axelor/aos-mobile-core';
-import {Button, useThemeColor} from '@axelor/aos-mobile-ui';
-
-const TaskButton = () => {
-  const Colors = useThemeColor();
-  const I18n = useTranslator();
-
-  return (
-    <Button
-      color={Colors.progressColor}
-      title={I18n.t('Project_LogTime')}
-      iconName="clock-history"
-    />
-  );
-};
-
-export default TaskButton;
