+++ packages/apps/project/src/screens/TaskDetailsScreen.tsx
@@ -16,83 +16,25 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, {useCallback, useEffect} from 'react';
-import {StyleSheet} from 'react-native';
-import {HeaderContainer, Screen, ScrollView} from '@axelor/aos-mobile-ui';
-import {
-  CustomFieldForm,
-  useDispatch,
-  useSelector,
-} from '@axelor/aos-mobile-core';
-import {
-  ProjectSimpleCard,
-  EditTaskButton,
-  TaskButton,
-  TaskDetailsHeader,
-  TimeSpentGridView,
-  ProjectTaskDropdownCards,
-} from '../components';
-import {fetchProjectStatus} from '../features/projectSlice';
-import {fetchProjectTaskById} from '../features/projectTaskSlice';
+import React from 'react';
+import {BottomBar, Screen} from '@axelor/aos-mobile-ui';
+import {TaskGeneralInformationView} from '../components';
 
-const TaskDetailsScreen = ({navigation, route}) => {
+const TaskDetailsScreen = ({route}) => {
   const projecTaskId = route?.params?.projecTaskId;
-  const dispatch = useDispatch();
 
-  const {projectTask, loadingProjectTask} = useSelector(
-    (state: any) => state.project_projectTask,
-  );
-
-  const refresh = useCallback(() => {
-    dispatch((fetchProjectTaskById as any)({projecTaskId}));
-  }, [dispatch, projecTaskId]);
-
-  useEffect(() => {
-    dispatch(fetchProjectStatus());
-    refresh();
-  }, [dispatch, refresh]);
-
-  if (projecTaskId !== projectTask?.id) {
-    return null;
-  }
+  const bottomBarItems = [
+    {
+      iconName: 'house',
+      viewComponent: <TaskGeneralInformationView projecTaskId={projecTaskId} />,
+    },
+  ];
 
   return (
-    <Screen removeSpaceOnTop={true} fixedItems={<TaskButton />}>
-      <HeaderContainer
-        expandableFilter={false}
-        fixedItems={<TaskDetailsHeader />}
-      />
-      <ScrollView
-        style={styles.container}
-        refresh={{loading: loadingProjectTask, fetcher: refresh}}>
-        <ProjectSimpleCard
-          code={projectTask.project?.code}
-          name={projectTask.project?.name}
-          projectStatus={projectTask.project?.projectStatus}
-          onPress={() => {
-            navigation.navigate('ProjectDetailsScreen', {
-              projectId: projectTask.project?.id,
-            });
-          }}
-        />
-        <ProjectTaskDropdownCards />
-        <CustomFieldForm
-          model="com.axelor.apps.project.db.ProjectTask"
-          fieldType="attrs"
-          modelId={projectTask.id}
-          readonly
-        />
-        <TimeSpentGridView />
-      </ScrollView>
-      <EditTaskButton idTask={projectTask?.id} />
+    <Screen removeSpaceOnTop={true}>
+      <BottomBar items={bottomBarItems} />
     </Screen>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    height: null,
-  },
-});
 
 export default TaskDetailsScreen;
